// src/main/requests/types/UpdateCourtFacilitiesRequest.ts
export interface UpdateCourtFacilitiesRequest {
  parking: boolean;
  freeWaterDispensers: boolean;
  snackVendingMachines: boolean;
  drinkVendingMachines: boolean;
  cafeteria: boolean;
  waitingArea: boolean;
  waitingAreaChildren: boolean;
  quietRoom: boolean;
  babyChanging: boolean;
  wifi: boolean;
}